chore: 多项修复
- watchfiles 日志抑制（main.py）
- 语意识别规则序号校正 + 周期支持小数（intent_recognition.md.j2）
- BrandInput.period/budget 从 int 改为 float 兼容小数（chat.py）
- 投放周期从 select 改为 number 自由输入（ScreenBrief.tsx）
 - period 匹配用 match(/[\d.]+/) 修复 1.5→1.56 bug
- start-backend.sh 加 --reload-exclude 避免日志触发热重载

+++ backend/app/schemas/chat.py
@@ -5,8 +5,8 @@ class BrandInput(BaseModel):
     brand_name: str | None = Field(None, description="品牌名称")
     category: str | None = Field(None, description="品牌品类，英文 snake_case")
     city: str | None = Field(None, description="目标城市")
-    budget: int | None = Field(None, ge=1, description="预算，单位万元")
-    period: int | None = Field(None, ge=1, description="周期，单位月")
+    budget: float | None = Field(None, ge=1, description="预算，单位万元")
+    period: float | None = Field(None, ge=1, description="周期，单位月")
 
 
 class ChatRequest(BaseModel):
